Refactor: 패키지 구조 추가 수정
Before: 240815_01:14

Note: 경로 수정 및 Class 이름 변경 (ContractService.java => DonateWeb3jService.java)
