+++ src/main/java/com/hikdonate/donate/donor/service/DonateWeb3jService.java
@@ -1,4 +1,4 @@
-package com.hikdonate.donate.web3jAPI;
+package com.hikdonate.donate.donor.service;
 
 import com.hikdonate.donate.web3jAPI.contracts.DonateTokenBank;
 import com.hikdonate.donate.web3jAPI.contracts.MultiTokenTransfer;
@@ -7,13 +7,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
-public class ContractService {
+public class DonateWeb3jService {
     private DonateTokenBank donateTokenBank;
     private TokenTransfer tokenTransfer;
     private MultiTokenTransfer multiTokenTransfer;
 
     @Autowired
-    public ContractService(DonateTokenBank donateTokenBank, TokenTransfer tokenTransfer, MultiTokenTransfer multiTokenTransfer) {
+    public DonateWeb3jService(DonateTokenBank donateTokenBank, TokenTransfer tokenTransfer, MultiTokenTransfer multiTokenTransfer) {
         this.donateTokenBank = donateTokenBank;
         this.tokenTransfer = tokenTransfer;
         this.multiTokenTransfer = multiTokenTransfer;
