+++ src/main/java/com/hikdonate/donate/transaction/repository/TransactionRepository.java
@@ -1,4 +1,4 @@
-package com.hikdonate.donate.admin.repository;
+package com.hikdonate.donate.transaction.repository;
 
 import com.hikdonate.donate.transaction.Transaction;
 import org.springframework.data.mongodb.repository.MongoRepository;
